rename modular to convention
