+++ features/entrypoint/build.gradle.kts
@@ -2,7 +2,7 @@ apply<plugin.Junit5Plugin>()
 apply<plugin.spotless.SpotlessPlugin>()
 
 plugins {
-    id("com.modular.composelib")
+    id("com.convention.composeandroidlib")
 }
 dependencies {
     implementation(projects.platform.designSystem)
