+++ features/recipe/build.gradle.kts
@@ -2,7 +2,7 @@ apply<plugin.Junit5Plugin>()
 apply<plugin.spotless.SpotlessPlugin>()
 @Suppress("DSL_SCOPE_VIOLATION")
 plugins {
-    id("com.modular.composelib")
+    id("com.convention.composeandroidlib")
     alias(libs.plugins.ksp)
 }
