+++ pluginBuild/convention/src/main/kotlin/com/convention/ComposeAndroidLib.kt
@@ -1,13 +1,12 @@
-package com.modular.plugin
+package com.convention
 
 import com.android.build.gradle.LibraryExtension
-import com.modular.plugin.configs.Dependencies
-import com.modular.plugin.extensions.setupCompose
-import com.modular.plugin.extensions.setupSdk
+import com.convention.configs.Dependencies
+import com.convention.extensions.setupCompose
+import com.convention.extensions.setupSdk
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
-import org.gradle.kotlin.dsl.kotlin
 
 class ComposeAndroidLib: Plugin<Project> {
     override fun apply(project: Project) {
